staging android: Keep timestamp info of merged fences.
Before this patch, sync_fence_merge discarded sync points that
already signaled. This is adequate for the purposes of meeting
ordering dependencies and event notification. However, it does
not allow for proper retrieval of the final event timestamp
via sync_fence_info.

It is particularly troublesome when merging fences where all
the sync points have already signaled and all timestamp info
gets discarded.

This patch adds sync points even if they are retired and
updates the logic that initializes the fence->status to the
number of unsignaled sync points.

This behavior matches the behavior before the fence api
conversion in:
commit 0f0d8406fb9c ("android: convert sync to fence api, v6")

Devices based on kernels prior to 3.17 do not have this bug.

I verified his fixes eglGetFrameTimestampsANDROID on devices
that rely on newer versions of the kernel.

Bug: 66916041
Signed-off-by: Brian Anderson <brianderson@google.com>
Change-Id: Ia4c48029225979413701b9c97d6a76041715a304
Signed-off-by: Ratoriku <a1063021545@gmail.com>